+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Dev
 - Unminified idents are now made by css-loader's defaultGetLocalIdent
+- New "enabled" option to toggle the plug-in, by default true on production
 
 # Version 0.3.0
 - Use "compilation" and "afterProcessAssets" hooks instead of "afterEmit"